backend/requirements.txt
python-dotenv
pydantic-settings
sqlalchemy
mariadb
pymysql
cryptography
